requirements.txt
elasticsearch>=8.0.0
requests>=2.28.0
boto3>=1.26.0
pyyaml>=6.0
slack-sdk>=3.19.0
